+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+
+## [2.19.0] - 2025-10-16
 ### Added
 - CashAmount, CashierBankCode, CashierType, Description and KeyId properties to BrcodePreview
 - webhook event parser for merchant and corporate resources